Refactor lifetimeStats APIFields models
