+++ src/client/lib/authAPI/index.ts
@@ -1,6 +1,6 @@
 import { getAPIWithAuth } from '../fetchAPI'
 
-import { SHARED_APIFields__LifetimeStats } from '@/shared/models/apiFields'
+import { SHARED_APIFields__LifetimeStats } from '@/shared/models/apiFields/lifetimeStats'
 import { SHARED_APIFields__GetUserCard } from '@/shared/models/apiFields/userCard'
 
 /*
