+++ src/shared/models/models/Stats.ts
@@ -1,6 +1,3 @@
-import { SHARED_Model__AllLanguageStats } from './Language'
-import { SHARED_Model__RepoWithCommitCountsAndLanguages } from './Repos'
-
 export const makeLinesStats = (): SHARED_Model__LinesStats => {
     return {
         numAdditions: 0,
@@ -18,10 +15,4 @@ export interface SHARED_Model__LinesStats {
 export interface SHARED_Model__RepoCommitCountStats {
     numRepos: number
     numCommits: number
-}
-
-export interface SHARED_Model__LifetimeStats {
-    language_stats: SHARED_Model__AllLanguageStats
-    repos: SHARED_Model__RepoWithCommitCountsAndLanguages[]
-    rc_stats: SHARED_Model__RepoCommitCountStats
 }
